requirements.txt
fastapi>=0.115
uvicorn>=0.29
gspread>=6.0
google-auth>=2.0
google-auth-oauthlib>=1.0
requests>=2.31
python-multipart>=0.0.9
playwright>=1.40
